refactor(personality-insights-v3): remove compatibility layer for personality insights v3
BREAKING CHANGE: Deprecated methods in Personality Insights v3 are no longer available. Changed parameter names are no longer interally corrected.

To migrate your code, use the methods and parameters currently available with the service as documented here: https://www.ibm.com/watson/developercloud/personality-insights/api/v3/node.html?node
