+++ examples/personality_insights.v3.js
@@ -78,7 +78,7 @@ personalityInsights.profile(
  * https://console.bluemix.net/docs/services/personality-insights/output-csv.html#outputCSV
  */
 personalityInsights
-  .profile_csv({
+  .profileAsCsv({
     content: 'Enter more than 100 unique words here...',
     content_type: 'text/plain',
     csv_headers: true
